refactor!: move `FormCombobox` into `Form` namespace
